move functions outside class to db helpers
