+++ packages/nocodb/src/helpers/dbHelpers.ts
@@ -1,9 +1,20 @@
-import { type NcContext, RelationTypes } from 'nocodb-sdk';
+import {
+  isCreatedOrLastModifiedByCol,
+  isCreatedOrLastModifiedTimeCol,
+  isOrderCol,
+  type NcContext,
+  RelationTypes,
+  UITypes,
+} from 'nocodb-sdk';
+import { customAlphabet } from 'nanoid';
+import { v4 as uuidv4 } from 'uuid';
+import type { BaseModelSqlv2 } from 'src/db/BaseModelSqlv2';
 import type CustomKnex from 'src/db/CustomKnex';
 import type { Knex } from 'knex';
+import type { Filter } from '~/models';
 import NcConnectionMgrv2 from '~/utils/common/NcConnectionMgrv2';
 import {
-  type Column,
+  Column,
   type LinkToAnotherRecordColumn,
   Model,
   Source,
@@ -176,4 +187,124 @@ export function getRelatedLinksColumn(
 
 export function extractIdPropIfObjectOrReturn(id: any, prop: string) {
   return typeof id === 'object' ? id[prop] : id;
+}
+export const nanoidv2 = customAlphabet(
+  '1234567890abcdefghijklmnopqrstuvwxyz',
+  14,
+);
+
+export async function populatePk(
+  context: NcContext,
+  model: Model,
+  insertObj: any,
+) {
+  await model.getColumns(context);
+  for (const pkCol of model.primaryKeys) {
+    if (!pkCol.meta?.ag || insertObj[pkCol.title]) continue;
+    insertObj[pkCol.title] =
+      pkCol.meta?.ag === 'nc' ? `rc_${nanoidv2()}` : uuidv4();
+  }
+}
+
+export function checkColumnRequired(
+  column: Column<any>,
+  fields: string[],
+  extractPkAndPv?: boolean,
+) {
+  // if primary key or foreign key included in fields, it's required
+  if (column.pk || column.uidt === UITypes.ForeignKey) return true;
+
+  if (extractPkAndPv && column.pv) return true;
+
+  // check fields defined and if not, then select all
+  // if defined check if it is in the fields
+  return !fields || fields.includes(column.title);
+}
+
+export async function getColumnName(
+  context: NcContext,
+  column: Column<any>,
+  columns?: Column[],
+) {
+  if (
+    !isCreatedOrLastModifiedTimeCol(column) &&
+    !isCreatedOrLastModifiedByCol(column) &&
+    !isOrderCol(column)
+  )
+    return column.column_name;
+  columns =
+    columns ||
+    (await Column.list(context, { fk_model_id: column.fk_model_id }));
+
+  switch (column.uidt) {
+    case UITypes.CreatedTime: {
+      const createdTimeSystemCol = columns.find(
+        (col) => col.system && col.uidt === UITypes.CreatedTime,
+      );
+      if (createdTimeSystemCol) return createdTimeSystemCol.column_name;
+      return column.column_name || 'created_at';
+    }
+    case UITypes.LastModifiedTime: {
+      const lastModifiedTimeSystemCol = columns.find(
+        (col) => col.system && col.uidt === UITypes.LastModifiedTime,
+      );
+      if (lastModifiedTimeSystemCol)
+        return lastModifiedTimeSystemCol.column_name;
+      return column.column_name || 'updated_at';
+    }
+    case UITypes.CreatedBy: {
+      const createdBySystemCol = columns.find(
+        (col) => col.system && col.uidt === UITypes.CreatedBy,
+      );
+      if (createdBySystemCol) return createdBySystemCol.column_name;
+      return column.column_name || 'created_by';
+    }
+    case UITypes.LastModifiedBy: {
+      const lastModifiedBySystemCol = columns.find(
+        (col) => col.system && col.uidt === UITypes.LastModifiedBy,
+      );
+      if (lastModifiedBySystemCol) return lastModifiedBySystemCol.column_name;
+      return column.column_name || 'updated_by';
+    }
+    case UITypes.Order: {
+      const orderSystemCol = columns.find(
+        (col) => col.system && col.uidt === UITypes.Order,
+      );
+      if (orderSystemCol) return orderSystemCol.column_name;
+      return column.column_name || 'nc_order';
+    }
+    default:
+      return column.column_name;
+  }
+}
+
+export function getAs(column: Column) {
+  return column.asId || column.id;
+}
+
+export function replaceDynamicFieldWithValue(
+  _row: any,
+  _rowId,
+  _tableColumns: Column[],
+  _readByPk: typeof BaseModelSqlv2.prototype.readByPk,
+  _queryParams?: Record<string, string>,
+) {
+  const replaceWithValue = async (conditions: Filter[]) => {
+    return conditions;
+  };
+  return replaceWithValue;
+}
+
+export const isPrimitiveType = (val) =>
+  typeof val === 'string' || typeof val === 'number';
+
+export function transformObject(value, idToAliasMap) {
+  const result = {};
+  Object.entries(value).forEach(([k, v]) => {
+    const btAlias = idToAliasMap[k];
+    if (btAlias) {
+      result[btAlias] = v;
+    }
+  });
+  return result;
 }
